fix: blocks field must be short and mixed with add blocks tag

+++ src/main/java/me/oczi/schematic/Schematic.java
@@ -37,7 +37,7 @@ public interface Schematic {
 
     short getHeight();
 
-    byte[] getBlocks();
+    short[] getBlocks();
 
     byte[] getData();
